SWIDEA_SITE/requirements.txt
django-bootstrap-v5==1.0.8